add usage_log_t template class to simplify trace-logging
if the customer requirements are such that you must comprensively
log all calls into a particular API (and that API is represented
by a pointer you hold), then this template class makes it easy
to log all such calls, and to do so compactly on the same line
as each call. Doing it 'inline' like this makes it easier to
always remember to log, and it makes it essentially impossible
to remove an API call and forget to remove the associated log.